Add Trend#create_or_update_observations helper
The react form-component that deals with the trend data doesn't handle
the hidden `_destroy` fields. While this solution isn't exactly pretty,
I think it's ok for the time being as a) the number of data points is
usually small and b) I think in the short-mid term we'll migrate away
from having a dedicated TrendObservation and inline the data via
a JSON/Array field on the Trend model.